Added functional helpers
The idea is that these will make sequential model updates easier and
type safe. Using functions is great but JS doesn't give you a lot of
sugar to make it easy. If you want to apply a bunch of operations on a
model it will look something like this in JS:

        a(a1, b(c(c1, model)));

These utils will turn it into this:

        call(chain3(
            arg1(a, a1),
            arg(b),
            arg1(c, c1)
        ), model)

Or, for convenience,

        call3(
            arg1(a, a1),
            arg(b),
            arg1(c, c1),
            model)

Or, if all of the functions take the same input and output:

        apply([
            arg1(a, a1),
            arg(b),
            arg1(c, c1)
        ], model)