Removed modelSet
I wanted to include this as a convenience but apparently typescript
disables some of the contextual lookup behavior if a function signature
includes generics. This results in the following being legal:

        arg2(modelSet, 'nonExistingKey', arg)

The solution so far is to not have  generic function signatures, which
means the application has to define a modelSet because only it knows
about the type of its model.